// web/src/api/models/createRelationshipRequestStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CreateRelationshipRequestStatus = typeof CreateRelationshipRequestStatus[keyof typeof CreateRelationshipRequestStatus];


export const CreateRelationshipRequestStatus = {
    MARRIED: 'MARRIED',
    PARTNERED: 'PARTNERED',
    SEPARATED: 'SEPARATED',
    DIVORCED: 'DIVORCED',
    WIDOWED: 'WIDOWED',
} as const;