// web/src/api/models/createRelationshipRequestType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CreateRelationshipRequestType = typeof CreateRelationshipRequestType[keyof typeof CreateRelationshipRequestType];


export const CreateRelationshipRequestType = {
    PARENT_OF: 'PARENT_OF',
    SPOUSE_OF: 'SPOUSE_OF',
    PARTNER_OF: 'PARTNER_OF',
} as const;